Add Storage class to handle loading and saving of tasks in the file
