+++ data/pecky.txt
@@ -8,3 +8,4 @@ D|0|return book 3|2003-08-03 1000
 D|0|5|2004-05-06 0000
 E|1|testEvent|2003-04-05 0000|2003-05-06 0000
 E|0|testEvent2|2005-02-01 0000|2006-03-02 0000
+E|0|testEvent3|2007-06-05 0000|2009-08-07 0000
