+++ src/main/java/Ui.java
@@ -5,6 +5,7 @@ public class Ui {
     private static String HELLO = "Hello! I'm Pecky!\n" +
             "What can I do for you?";
     private static String BYE = "Bye. Hope to see you again soon!";
+    private static String UNKNOWN = "OOPS!!! I'm sorry, but I don't know what that means :-(";
     private static Scanner scanner = new Scanner(System.in);
 
     public static String getInput() {
@@ -23,5 +24,9 @@ public class Ui {
 
     public static void hello() {
         Ui.print(HELLO);
+    }
+
+    public static void unknown() {
+        Ui.print(UNKNOWN);
     }
 }
